requirements.txt
streamlit>=1.28.0
google-genai>=1.21.1
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
feedparser>=6.0.10